Adding a reminder TODO comment, about Starter selection.
Starter selection is currently greedy: it walks the given set of Slots,
and chooses the best available (unchosen, un-injured) player for each
Slot.

However, this doesn't always produce the best scoring set of Starters,
since it depends on the order in which the Slots are given.

This commit adds a to-do comment, about fixing this problem in the
future...